refactor(inventory): extract reusable UI components and externalize strings
This commit overhauls the UI architecture of the inventory feature by decomposing monolithic screen files into a library of atomic components. It improves maintainability through a dedicated `components` package and implements localization by migrating hardcoded strings to the resources catalog.

- **`ui/components`**:
    - Introduced a comprehensive suite of reusable components including `AnalyticsStatCard`, `WardrobeCard`, `MetricItem`, `ProInsightCard`, and `RankingStatCard`.
    - Consolidated duplicated layout logic for efficiency rows and ranking displays into standalone composables.
    - Implemented `AtelierWardrobeCard` for high-level category summaries with progress tracking and metadata display.
- **`WardrobeDetailScreen.kt`**:
    - Redesigned the garment detail view with a focus on editorial typography and a structured hero section.
    - Integrated new components for performance metrics (Investment, Cost/Wear, Usage) and composition details.
- **`WardrobeAnalyticsScreen.kt`**:
    - Refactored the "Style Intelligence" dashboard to use specialized stat cards for portfolio performance and usage leaderboards.
    - Standardized the visual presentation of the wardrobe palette and efficiency analysis.
- **`WardrobeEditScreen.kt`**:
    - Cleaned up the garment editing form, replacing manual layout code with organized `OutlinedTextField` rows and standardized save actions.
- **`WardrobeLandingScreen.kt`**:
    - Updated the "Style Archive" landing page with a modern, card-based summary of total pieces and portfolio value.
- **`strings.xml`**:
    - Added comprehensive string resources for titles, labels, descriptions, and empty-state prompts across all inventory screens.
- **General**:
    - Switched to `CenterAlignedTopAppBar` for improved visual hierarchy in primary feature views.
    - Standardized theme usage by leveraging `MaterialTheme` color schemes and typography tokens consistently.
