+++ applications/kocolor/features/inventory/src/main/java/com/zoewave/probase/kocolor/features/inventory/ui/ColorVerificationScreen.kt
@@ -1,27 +1,19 @@
 package com.zoewave.probase.kocolor.features.inventory.ui
 
-import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.*
 import androidx.compose.foundation.lazy.LazyColumn
 import androidx.compose.foundation.lazy.items
-import androidx.compose.foundation.shape.CircleShape
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.automirrored.filled.ArrowBack
 import androidx.compose.material3.*
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.getValue
-import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.draw.clip
-import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.layout.ContentScale
-import androidx.compose.ui.text.font.FontWeight
+import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
-import androidx.hilt.lifecycle.viewmodel.compose.hiltViewModel
-import androidx.lifecycle.compose.collectAsStateWithLifecycle
-import coil.compose.AsyncImage
-import com.zoewave.probase.kocolor.features.inventory.util.toComposeColor
+import com.zoewave.probase.kocolor.features.inventory.R
+import com.zoewave.probase.kocolor.features.inventory.ui.components.ColorVerificationItem
+import com.zoewave.probase.kocolor.model.ClothingCategory
 import com.zoewave.probase.kocolor.model.ClothingItem
 import com.zoewave.probase.kocolor.model.KoColorRoute
 
@@ -29,19 +21,6 @@ data class ColorVerificationUiState(
     val items: List<ClothingItem> = emptyList()
 )
 
-@Preview(showBackground = true)
-@Composable
-private fun ColorVerificationRoutePreview() {
-    MaterialTheme {
-        ColorVerificationRoute(
-            uiState = ColorVerificationUiState(),
-            onEvent = {},
-            navTo = {}
-        )
-    }
-}
-
-@OptIn(ExperimentalMaterial3Api::class)
 @Composable
 fun ColorVerificationRoute(
     uiState: ColorVerificationUiState,
@@ -61,7 +40,7 @@ private fun ColorVerificationScreenPreview() {
     MaterialTheme {
         ColorVerificationScreen(
             uiState = ColorVerificationUiState(
-                items = listOf(ClothingItem(name = "Item", category = com.zoewave.probase.kocolor.model.ClothingCategory.TOPS))
+                items = listOf(ClothingItem(name = "Item", category = ClothingCategory.TOPS))
             ),
             onEvent = {},
             navTo = {}
@@ -79,10 +58,10 @@ fun ColorVerificationScreen(
     Scaffold(
         topBar = {
             TopAppBar(
-                title = { Text("Color Verification") },
+                title = { Text(stringResource(R.string.applications_kocolor_features_inventory_test_colors)) },
                 navigationIcon = {
                     IconButton(onClick = { navTo(KoColorRoute.Back) }) {
-                        Icon(Icons.AutoMirrored.Filled.ArrowBack, contentDescription = "Back")
+                        Icon(Icons.AutoMirrored.Filled.ArrowBack, contentDescription = stringResource(R.string.applications_kocolor_features_inventory_back))
                     }
                 }
             )
@@ -97,112 +76,6 @@ fun ColorVerificationScreen(
         ) {
             items(uiState.items) { item ->
                 ColorVerificationItem(uiState = item, onEvent = {}, navTo = {})
-            }
-        }
-    }
-}
-
-@Preview(showBackground = true)
-@Composable
-private fun ColorVerificationItemPreview() {
-    MaterialTheme {
-        ColorVerificationItem(
-            uiState = ClothingItem(name = "Item", category = com.zoewave.probase.kocolor.model.ClothingCategory.TOPS),
-            onEvent = {},
-            navTo = {}
-        )
-    }
-}
-
-@Composable
-fun ColorVerificationItem(
-    uiState: ClothingItem,
-    onEvent: (Unit) -> Unit,
-    navTo: (KoColorRoute) -> Unit
-) {
-    val item = uiState
-    Card(
-        modifier = Modifier.fillMaxWidth(),
-        elevation = CardDefaults.cardElevation(defaultElevation = 2.dp)
-    ) {
-        Row(
-            modifier = Modifier
-                .padding(12.dp)
-                .height(IntrinsicSize.Min),
-            verticalAlignment = Alignment.CenterVertically
-        ) {
-            // Image
-            AsyncImage(
-                model = item.imageUrl,
-                contentDescription = null,
-                modifier = Modifier
-                    .size(80.dp)
-                    .clip(MaterialTheme.shapes.medium)
-                    .background(MaterialTheme.colorScheme.surfaceVariant),
-                contentScale = ContentScale.Crop
-            )
-
-            Spacer(modifier = Modifier.width(16.dp))
-
-            // Data
-            Column(
-                modifier = Modifier.weight(1f)
-            ) {
-                Text(
-                    text = item.name,
-                    style = MaterialTheme.typography.titleMedium,
-                    fontWeight = FontWeight.Bold
-                )
-                
-                Spacer(modifier = Modifier.height(4.dp))
-
-                Row(verticalAlignment = Alignment.CenterVertically) {
-                    Text(
-                        text = "Dominant: ${item.dominantHex ?: "N/A"}",
-                        style = MaterialTheme.typography.bodySmall
-                    )
-                    Spacer(modifier = Modifier.width(8.dp))
-                    item.dominantHex?.let {
-                        Box(
-                            modifier = Modifier
-                                .size(16.dp)
-                                .clip(CircleShape)
-                                .background(it.toComposeColor())
-                        )
-                    }
-                }
-
-                Spacer(modifier = Modifier.height(4.dp))
-
-                // Palette
-                Row(
-                    verticalAlignment = Alignment.CenterVertically,
-                    horizontalArrangement = Arrangement.spacedBy(4.dp)
-                ) {
-                    Text(
-                        text = "Palette: ",
-                        style = MaterialTheme.typography.bodySmall
-                    )
-                    item.paletteHexes.forEach { hex ->
-                        Box(
-                            modifier = Modifier
-                                .size(12.dp)
-                                .clip(CircleShape)
-                                .background(hex.toComposeColor())
-                        )
-                    }
-                }
-
-                Spacer(modifier = Modifier.height(4.dp))
-
-                Text(
-                    text = "Contrast: ${item.contrastLevel ?: "N/A"}",
-                    style = MaterialTheme.typography.bodySmall
-                )
-                Text(
-                    text = "Group: ${item.koColorGroup ?: "N/A"}",
-                    style = MaterialTheme.typography.bodySmall
-                )
             }
         }
     }
