+++ applications/kocolor/features/inventory/src/main/java/com/zoewave/probase/kocolor/features/inventory/ui/WardrobeDetailScreen.kt
@@ -1,8 +1,8 @@
 package com.zoewave.probase.kocolor.features.inventory.ui
 
+import androidx.compose.foundation.BorderStroke
 import androidx.compose.foundation.background
 import androidx.compose.foundation.border
-import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.*
 import androidx.compose.foundation.rememberScrollState
 import androidx.compose.foundation.shape.CircleShape
@@ -10,35 +10,33 @@ import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.foundation.verticalScroll
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.automirrored.filled.ArrowBack
-import androidx.compose.material.icons.automirrored.filled.MenuBook
-import androidx.compose.material.icons.automirrored.filled.TrendingDown
 import androidx.compose.material.icons.filled.*
 import androidx.compose.material3.*
-import androidx.compose.runtime.*
+import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.draw.alpha
 import androidx.compose.ui.draw.clip
-import androidx.compose.ui.graphics.Brush
 import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.graphics.vector.ImageVector
 import androidx.compose.ui.layout.ContentScale
+import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.text.font.FontFamily
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import coil.compose.AsyncImage
+import com.zoewave.probase.features.graphics.colorpicker.util.isColorDark
+import com.zoewave.probase.features.graphics.colorpicker.util.parseColor
+import com.zoewave.probase.kocolor.features.inventory.R
+import com.zoewave.probase.kocolor.features.inventory.ui.components.DetailRow
+import com.zoewave.probase.kocolor.features.inventory.ui.components.MetricItem
+import com.zoewave.probase.kocolor.features.inventory.ui.components.ProInsightCard
+import com.zoewave.probase.kocolor.features.inventory.ui.components.SectionHeader
+import com.zoewave.probase.kocolor.model.ClothingCategory
 import com.zoewave.probase.kocolor.model.ClothingItem
 import com.zoewave.probase.kocolor.model.KoColorRoute
-import com.zoewave.probase.features.graphics.colorpicker.util.parseColor
 import java.text.NumberFormat
-import java.util.Locale
-
-private fun isColorDark(color: Color): Boolean {
-    val luminance = 0.299 * color.red + 0.587 * color.green + 0.114 * color.blue
-    return luminance < 0.5
-}
+import java.util.*
 
 data class WardrobeDetailUiState(
     val itemId: Long,
@@ -51,10 +49,19 @@ private fun WardrobeDetailScreenPreview() {
     MaterialTheme {
         WardrobeDetailScreen(
             uiState = WardrobeDetailUiState(
-                itemId = 1L,
+                itemId = 1,
                 wardrobeUiState = WardrobeUiState(
                     items = listOf(
-                        ClothingItem(id = 1L, name = "Silk Blouse", category = com.zoewave.probase.kocolor.model.ClothingCategory.TOPS, price = 120.0)
+                        ClothingItem(
+                            id = 1, 
+                            name = "Silk Blouse", 
+                            brand = "Celine", 
+                            category = ClothingCategory.TOPS, 
+                            price = 850.0, 
+                            usageCount = 12, 
+                            colorHex = "#F5F5F5",
+                            material = "100% Silk"
+                        )
                     )
                 )
             ),
@@ -71,31 +78,21 @@ fun WardrobeDetailScreen(
     onEvent: (WardrobeEvent) -> Unit,
     navTo: (KoColorRoute) -> Unit
 ) {
-    val itemId = uiState.itemId
-    val state = uiState.wardrobeUiState
-    val item = remember(state.items, itemId) {
-        state.items.find { it.id == itemId }
-    } ?: return
-
-    val itemColor = item.dominantHex?.let { parseColor(it) } 
-        ?: item.colorHex?.let { parseColor(it) } 
-        ?: MaterialTheme.colorScheme.surfaceVariant
+    val item = uiState.wardrobeUiState.items.find { it.id == uiState.itemId } ?: return
+    val atelierBrown = Color(0xFF8B5E3C)
 
     Scaffold(
         topBar = {
-            TopAppBar(
-                title = { Text("Garment Details", style = MaterialTheme.typography.labelLarge) },
+            CenterAlignedTopAppBar(
+                title = { Text(stringResource(R.string.applications_kocolor_features_inventory_garment_details), style = MaterialTheme.typography.labelLarge) },
                 navigationIcon = {
                     IconButton(onClick = { navTo(KoColorRoute.Back) }) {
-                        Icon(Icons.AutoMirrored.Filled.ArrowBack, contentDescription = "Back")
+                        Icon(Icons.AutoMirrored.Filled.ArrowBack, contentDescription = stringResource(R.string.applications_kocolor_features_inventory_back))
                     }
                 },
                 actions = {
-                    IconButton(onClick = {
-                        onEvent(WardrobeEvent.UpdateDraft(item))
-                        navTo(KoColorRoute.WardrobeEdit(item.id))
-                    }) {
-                        Icon(Icons.Default.Edit, contentDescription = "Edit")
+                    IconButton(onClick = { navTo(KoColorRoute.WardrobeEdit(item.id)) }) {
+                        Icon(Icons.Default.Edit, contentDescription = stringResource(R.string.applications_kocolor_features_inventory_edit))
                     }
                 }
             )
@@ -106,185 +103,121 @@ fun WardrobeDetailScreen(
                 .padding(padding)
                 .fillMaxSize()
                 .verticalScroll(rememberScrollState())
+                .background(Color.White)
         ) {
-            // Hero
-            Box(
+            // 1. Hero Image Section
+            Surface(
                 modifier = Modifier
                     .fillMaxWidth()
-                    .height(400.dp)
+                    .aspectRatio(0.85f),
+                color = Color(0xFFFBF8F5)
             ) {
-                if (item.imageUrl != null) {
-                    Column(modifier = Modifier.fillMaxSize()) {
+                Box(modifier = Modifier.fillMaxSize()) {
+                    if (item.imageUrl != null) {
                         AsyncImage(
                             model = item.imageUrl,
                             contentDescription = item.name,
-                            modifier = Modifier.fillMaxWidth().weight(1f),
+                            modifier = Modifier.fillMaxSize(),
                             contentScale = ContentScale.Crop
                         )
-                        Box(modifier = Modifier.fillMaxWidth().height(120.dp).background(itemColor))
                     }
-                } else {
-                    Box(modifier = Modifier.fillMaxSize().background(itemColor))
-                }
-
-                // Scrim for readability
-                Box(
-                    modifier = Modifier
-                        .fillMaxSize()
-                        .background(
-                            Brush.verticalGradient(
-                                colors = listOf(Color.Transparent, Color.Black.copy(alpha = 0.4f)),
-                                startY = 600f
-                            )
+                    
+                    // Category Badge
+                    Surface(
+                        modifier = Modifier.padding(24.dp).align(Alignment.TopStart),
+                        color = Color.White.copy(alpha = 0.9f),
+                        shape = RoundedCornerShape(12.dp)
+                    ) {
+                        Text(
+                            text = item.category.name,
+                            modifier = Modifier.padding(horizontal = 12.dp, vertical = 6.dp),
+                            style = MaterialTheme.typography.labelSmall,
+                            fontWeight = FontWeight.Black
                         )
-                )
-                
-                Column(modifier = Modifier.align(Alignment.BottomStart).padding(24.dp)) {
-                    Text(
-                        text = (item.brand ?: "KOCOLOR").uppercase(),
-                        style = MaterialTheme.typography.labelMedium,
-                        color = Color.White.copy(alpha = 0.8f),
-                        fontWeight = FontWeight.Black,
-                        letterSpacing = 2.sp
-                    )
-                    Text(
-                        text = item.name,
-                        style = MaterialTheme.typography.displaySmall,
-                        color = Color.White,
-                        fontFamily = FontFamily.Serif,
-                        fontWeight = FontWeight.Bold
-                    )
-                    Spacer(Modifier.height(12.dp))
-                    Surface(color = Color.White.copy(alpha = 0.9f), shape = RoundedCornerShape(12.dp)) {
-                        Text(text = item.category.name, modifier = Modifier.padding(horizontal = 12.dp, vertical = 6.dp), style = MaterialTheme.typography.labelSmall, fontWeight = FontWeight.Black)
                     }
                 }
             }
 
-            // Metrics
-            Row(modifier = Modifier.padding(24.dp).fillMaxWidth(), horizontalArrangement = Arrangement.spacedBy(16.dp)) {
-                val currencyFormatter = NumberFormat.getCurrencyInstance(Locale.US)
-                MetricItem(
-                    uiState = Triple("INVESTMENT", item.price?.let { currencyFormatter.format(it) } ?: "---", Icons.Default.Payments),
-                    onEvent = {},
-                    navTo = {},
-                    modifier = Modifier.weight(1f)
-                )
-                MetricItem(
-                    uiState = Triple("COST / WEAR", item.costPerUse?.let { currencyFormatter.format(it) } ?: "---", Icons.AutoMirrored.Filled.TrendingDown),
-                    onEvent = {},
-                    navTo = {},
-                    modifier = Modifier.weight(1f)
-                )
-                MetricItem(
-                    uiState = Triple("WEARS", item.usageCount.toString(), Icons.Default.History),
-                    onEvent = {},
-                    navTo = {},
-                    modifier = Modifier.weight(1f)
-                )
-            }
+            // 2. Data Content
+            Column(modifier = Modifier.padding(24.dp)) {
+                Row(
+                    modifier = Modifier.fillMaxWidth(),
+                    horizontalArrangement = Arrangement.SpaceBetween,
+                    verticalAlignment = Alignment.Top
+                ) {
+                    Column(modifier = Modifier.weight(1f)) {
+                        Text(
+                            text = item.brand ?: stringResource(R.string.applications_kocolor_features_inventory_brand_default),
+                            style = MaterialTheme.typography.labelLarge,
+                            color = atelierBrown,
+                            fontWeight = FontWeight.Bold,
+                            letterSpacing = 1.sp
+                        )
+                        Text(
+                            text = item.name,
+                            style = MaterialTheme.typography.headlineLarge,
+                            fontFamily = FontFamily.Serif,
+                            fontWeight = FontWeight.Bold
+                        )
+                    }
+                    
+                    // Dominant Color Dot
+                    val dominantColor = item.dominantHex?.let { parseColor(it) } ?: item.colorHex?.let { parseColor(it) } ?: Color.Transparent
+                    Surface(
+                        modifier = Modifier.size(40.dp),
+                        shape = CircleShape,
+                        color = dominantColor,
+                        border = BorderStroke(1.dp, Color.Black.copy(alpha = 0.1f))
+                    ) {}
+                }
 
-            // Wear Action
-            Button(
-                onClick = { onEvent(WardrobeEvent.WearItem(item.id)) },
-                modifier = Modifier.padding(horizontal = 24.dp).fillMaxWidth().height(56.dp),
-                shape = RoundedCornerShape(16.dp)
-            ) {
-                Icon(Icons.Default.Checkroom, null)
-                Spacer(Modifier.width(12.dp))
-                Text("Log Today's Wear", fontWeight = FontWeight.Bold)
-            }
+                Spacer(Modifier.height(32.dp))
 
-            Spacer(Modifier.height(32.dp))
+                // Performance Metrics
+                Row(
+                    modifier = Modifier.fillMaxWidth(),
+                    horizontalArrangement = Arrangement.SpaceBetween
+                ) {
+                    val currencyFormatter = NumberFormat.getCurrencyInstance(Locale.US)
+                    MetricItem(label = stringResource(R.string.applications_kocolor_features_inventory_investment), value = currencyFormatter.format(item.price ?: 0.0), modifier = Modifier.weight(1f))
+                    MetricItem(label = stringResource(R.string.applications_kocolor_features_inventory_cost_per_wear), value = item.costPerUse?.let { currencyFormatter.format(it) } ?: stringResource(R.string.applications_kocolor_features_inventory_not_available), modifier = Modifier.weight(1f))
+                    MetricItem(label = stringResource(R.string.applications_kocolor_features_inventory_wears), value = item.usageCount.toString(), modifier = Modifier.weight(1f))
+                }
 
-            // Detail Content
-            Column(modifier = Modifier.padding(horizontal = 24.dp), verticalArrangement = Arrangement.spacedBy(32.dp)) {
+                Spacer(Modifier.height(32.dp))
+
+                Button(
+                    onClick = { onEvent(WardrobeEvent.WearItem(item.id)) },
+                    modifier = Modifier.fillMaxWidth().height(56.dp),
+                    shape = RoundedCornerShape(12.dp),
+                    colors = ButtonDefaults.buttonColors(containerColor = atelierBrown)
+                ) {
+                    Icon(Icons.Default.Checkroom, null)
+                    Spacer(Modifier.width(8.dp))
+                    Text(stringResource(R.string.applications_kocolor_features_inventory_log_wear), fontWeight = FontWeight.Bold)
+                }
+
+                Spacer(Modifier.height(40.dp))
+
+                SectionHeader(stringResource(R.string.applications_kocolor_features_inventory_composition))
+                DetailRow(label = stringResource(R.string.applications_kocolor_features_inventory_material), value = item.material ?: stringResource(R.string.applications_kocolor_features_inventory_unknown))
                 
-                Column {
-                    SectionHeader(uiState = "Composition", onEvent = {}, navTo = {})
-                    DetailRow(uiState = "Material" to (item.material ?: "Unknown"), onEvent = {}, navTo = {})
-                }
-                Column {
-                    SectionHeader(uiState = "Archive Notes", onEvent = {}, navTo = {})
-                    ProInsightCard(uiState = item.notes to Icons.AutoMirrored.Filled.MenuBook, onEvent = {}, navTo = {})
-                }
+                Spacer(Modifier.height(24.dp))
+                
+                SectionHeader(stringResource(R.string.applications_kocolor_features_inventory_archive_notes))
+                Text(
+                    text = item.notes ?: stringResource(R.string.applications_kocolor_features_inventory_no_notes),
+                    style = MaterialTheme.typography.bodyMedium,
+                    lineHeight = 24.sp,
+                    color = MaterialTheme.colorScheme.onSurfaceVariant
+                )
+
+                Spacer(Modifier.height(32.dp))
+                
+                ProInsightCard(text = "Strategic curated wardrobe collection.")
             }
-            Spacer(Modifier.height(100.dp))
-        }
-    }
-}
-
-@Preview(showBackground = true)
-@Composable
-private fun SectionHeaderPreview() {
-    MaterialTheme {
-        SectionHeader(uiState = "Composition", onEvent = {}, navTo = {})
-    }
-}
-
-@Composable
-private fun SectionHeader(uiState: String, onEvent: (Unit) -> Unit, navTo: (KoColorRoute) -> Unit) {
-    Text(text = uiState.uppercase(), style = MaterialTheme.typography.labelMedium, fontWeight = FontWeight.Black, letterSpacing = 2.sp, color = MaterialTheme.colorScheme.primary.copy(alpha = 0.8f))
-}
-
-@Preview(showBackground = true)
-@Composable
-private fun MetricItemPreview() {
-    MaterialTheme {
-        MetricItem(uiState = Triple("INVESTMENT", "$100", Icons.Default.Payments), onEvent = {}, navTo = {})
-    }
-}
-
-@Composable
-private fun MetricItem(uiState: Triple<String, String, ImageVector>, onEvent: (Unit) -> Unit, navTo: (KoColorRoute) -> Unit, modifier: Modifier = Modifier) {
-    val (label, value, icon) = uiState
-    Column(modifier = modifier) {
-        Icon(icon, null, modifier = Modifier.size(20.dp), tint = MaterialTheme.colorScheme.primary)
-        Spacer(Modifier.height(8.dp))
-        Text(text = value, style = MaterialTheme.typography.titleLarge, fontWeight = FontWeight.Bold)
-        Text(text = label, style = MaterialTheme.typography.labelSmall, color = MaterialTheme.colorScheme.onSurfaceVariant)
-    }
-}
-
-@Preview(showBackground = true)
-@Composable
-private fun DetailRowPreview() {
-    MaterialTheme {
-        DetailRow(uiState = "Material" to "Silk", onEvent = {}, navTo = {})
-    }
-}
-
-@Composable
-private fun DetailRow(uiState: Pair<String, String>, onEvent: (Unit) -> Unit, navTo: (KoColorRoute) -> Unit) {
-    val (label, value) = uiState
-    Row(modifier = Modifier.fillMaxWidth().padding(vertical = 12.dp), horizontalArrangement = Arrangement.SpaceBetween) {
-        Text(text = label, style = MaterialTheme.typography.bodyMedium, color = MaterialTheme.colorScheme.onSurfaceVariant)
-        Text(text = value, style = MaterialTheme.typography.bodyLarge, fontWeight = FontWeight.SemiBold)
-    }
-}
-
-@Preview(showBackground = true)
-@Composable
-private fun ProInsightCardPreview() {
-    MaterialTheme {
-        ProInsightCard(uiState = "Notes" to Icons.AutoMirrored.Filled.MenuBook, onEvent = {}, navTo = {})
-    }
-}
-
-@Composable
-private fun ProInsightCard(uiState: Pair<String?, ImageVector>, onEvent: (Unit) -> Unit, navTo: (KoColorRoute) -> Unit) {
-    val content = uiState.first
-    val icon = uiState.second
-    val isAvailable = !content.isNullOrBlank()
-    Surface(
-        modifier = Modifier.fillMaxWidth(),
-        color = MaterialTheme.colorScheme.surfaceVariant.copy(alpha = if (isAvailable) 0.5f else 0.1f),
-        shape = RoundedCornerShape(24.dp)
-    ) {
-        Column(modifier = Modifier.padding(20.dp)) {
-            Icon(icon, null, modifier = Modifier.size(18.dp), tint = MaterialTheme.colorScheme.primary)
-            Spacer(Modifier.height(12.dp))
-            Text(text = content ?: "No notes provided.", style = MaterialTheme.typography.bodyMedium, color = MaterialTheme.colorScheme.onSurface.copy(alpha = if (isAvailable) 0.8f else 0.2f), lineHeight = 22.sp)
+            
+            Spacer(modifier = Modifier.height(100.dp))
         }
     }
 }
