+++ applications/kocolor/features/inventory/src/main/res/values/strings.xml
@@ -49,4 +49,13 @@
     <string name="applications_kocolor_features_inventory_material_composition">Material Composition</string>
     <string name="applications_kocolor_features_inventory_curators_notes">Curator\'s Notes</string>
     <string name="applications_kocolor_features_inventory_not_available">N/A</string>
+
+    <string name="applications_kocolor_features_inventory_style_archive_title">Style Archive</string>
+    <string name="applications_kocolor_features_inventory_curated_closet_title">Curated Closet.</string>
+    <string name="applications_kocolor_features_inventory_closet_desc">A professional look at style investments.</string>
+    <string name="applications_kocolor_features_inventory_total_pieces_label">TOTAL PIECES</string>
+    <string name="applications_kocolor_features_inventory_total_value_label">TOTAL VALUE</string>
+    <string name="applications_kocolor_features_inventory_verticals_label">VERTICALS</string>
+    <string name="applications_kocolor_features_inventory_recently_added">RECENTLY ADDED</string>
+    <string name="applications_kocolor_features_inventory_see_all">See All</string>
 </resources>
